Update Bubble Sort.cpp
About Bubble Sort, bubble sort's working principle, algorithm analysis, best-worst case, such as information has been added. A person entering this pull request can easily learn about bubble sort.
In addition, information about the swap process used in the algorithm is also included in the code how to work. To make this process is fully explained in the lower example is explained in the sample. In some places, design changes were made to make it look pleasing to the eye.