Create minimum_cost_path.cpp
A cost matrix will be given by the user. A program needs to be developed to find the minimum cost path from (0,0) position to (m,n) position i.e from the initial element to the final element in a given array. The path must be chosen such a way that total cost of traversing from initial position to the final position should be the least.