Create O(n)_rotation_of_vector\array
This algorithm implements right rotation of array in O(n) time complexity it employs the following method-
Lets place place every number of the array at its desired correct position. But while doing this, we will lose the original elements. Thus, we need to store the number being replaced in a 'temp' variable. Then, we can place the replaced number(which is stored in temp) at its correct position and so on, n times, where n is the length of array. We have chosen n to be the number of replacements since we have to shift all the elements of the array(which is n). But, there could be a problem with this method, if n % k =0 where k = k%n (since a value of k larger than n eventually leads to a k equivalent to k%n). In this case, while picking up numbers to be placed at the correct position, we will eventually reach the number from which we originally started. Thus, in such a case, when we hit the original number's index again, we start the same process with the index following our starting point.