Create activity_selection.cpp
The activity-selection problem is a classic example of a greedy algorithm. The problem of scheduling several competing activities that require exclusive use of a common resource, with a goal of selecting a maximum number of mutually compatible activities can be solved by taking the activities in order of their finishing times. This is a C++14 implementation of the problem.