Added a function for creating a BST from an Array.
Created a new function block at line 31 named "createBST" that takes two parameter, one is the array (either sorted or unsorted) and another is the size of the array.
This function can be called upon choice of "7". where person will be asked for entering the size and then entering the array data accordingly, function will create a BST using the array data and return the root address in predefined root variable which is on line no. 117, I returned it in the predefined root variable so that people can perform other actions from their menu on the tree created using array. Although we can also do the same in a separate variable too. Other than these I have used few variables like "SizeIn of Int" and "DataSet of Int Array" on line 176 and also added a macro just below the header file for defining the max size of array, which is 100 for now but can be changed, I have tested my algorithm upon 10000 datasets.

I hope you like my contribution!