Use a pre-loaded vector for sorting.
Previously, we asked the user to enter the size of the vector and
populate it with an item at a time. This approach takes time to complete
and prone to errors

If the user entered say 'j', whiles populating the vector, std::cin fails
and the other values in the vector will remain zero as the program
continues.