Second red black tree program
I think my variation has helpful additional documentation and naming for functions/variables to make it clearer how different parts of a red black tree's source code work. I started the program as a college homework assignment but all of the code and implementation methods are my own. I'm happy to update or fix anything with my code if needed!