this comit is for Hackoctoberfest2023
Key Characteristics:

Push Operation: Elements are added to the top of the stack using the "push" operation. This operation places a new node at the head of the linked list.

Pop Operation: The "pop" operation removes the top element from the stack. It returns the data from the node and adjusts the pointer to the next node as the new top.

Peek Operation: The "peek" operation allows you to view the top element without removing it from the stack.

Empty Stack: An empty stack has no nodes in the linked list, and its top points to null or nullptr depending on the programming language.

Full Stack: In a linked list-based stack, it's typically not constrained by a fixed size, making it dynamically resizable as needed.

Common Use Cases: Linked list-based stacks are used in various algorithms and applications, such as parsing expressions, tracking function calls in recursion, and managing undo/redo functionality in software applications.

Time Complexity: The "push" and "pop" operations typically have a time complexity of O(1) because they involve simple adjustments of the pointers.