LongestCommon Subsequense
This code is for finding the lcs between 2 strings

A "Longest Common Subsequence" (LCS) is the longest sequence of characters that can be found in two or more strings, without necessarily being consecutive. It is a way to measure the similarity or matching between strings. The LCS doesn't require characters to appear in the same order as in the original strings. Instead, it focuses on finding the longest sequence of characters that are present in the same relative order in the given strings. It is a common problem in computer science and is often used in fields like data comparison, DNA sequence 
alignment, and text comparison.